refactor(ManageSSHKey): use new SSHKeyDetail component
