+++ src/components/common/SSHKeyDetail/types.tsx
@@ -1,3 +1,4 @@
 export type SSHKeyDetailProps = {
   sshKeyId: string
+  showDelete?: boolean
 }
